requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning & Deep Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
# Alternative: torch>=2.0.0  # Uncomment if using PyTorch instead

# Technical Analysis
ta>=0.11.0  # pandas-ta unavailable in this environment; ta provides core indicators

# Probabilistic Models
hmmlearn>=0.3.0
pgmpy>=0.1.19

# Data Sources & APIs
alpaca-trade-api>=3.0.0
alpaca-py>=0.21.0
yfinance>=0.2.28
requests>=2.31.0

# Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0  # Required for transformers/FinBERT

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Dashboard
dash>=2.11.0
plotly>=5.15.0
dash-bootstrap-components>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Data Storage
pyarrow>=12.0.0  # For Parquet files
openpyxl>=3.1.0  # For Excel files if needed

# Time Series
statsmodels>=0.14.0  # For ARIMA baseline

